Updated to UVVM v2 version 2024.03.08 - Please see CHANGES.TXT for details.

+++ bitvis_vip_i2c/script/maintenance_script/test.py
@@ -7,27 +7,34 @@ from itertools import product
 try:
     from hdlregression import HDLRegression
 except:
-    print('Unable to import HDLRegression module. See HDLRegression documentation for installation instructions.')
+    print(
+        "Unable to import HDLRegression module. See HDLRegression documentation for installation instructions."
+    )
     sys.exit(1)
 
 
-def cleanup(msg='Cleaning up...'):
+def cleanup(msg="Cleaning up..."):
     print(msg)
 
     sim_path = os.getcwd()
 
-    for files in os.listdir(sim_path):
-        path = os.path.join(sim_path, files)
-        try:
-            shutil.rmtree(path)
-        except:
-            os.remove(path)
+    # Check if the current directory is 'sim'
+    if os.path.basename(sim_path) == "sim":
+        for files in os.listdir(sim_path):
+            path = os.path.join(sim_path, files)
+            try:
+                shutil.rmtree(path)
+            except:
+                os.remove(path)
+    else:
+        print('Current directory is not "sim". Skipping cleanup.')
 
-print('Verify Bitvis VIP I2C')
 
-cleanup('Removing any previous runs.')
+print("Verify Bitvis VIP I2C")
 
-hr = HDLRegression(simulator='modelsim')
+cleanup("Removing any previous runs.")
+
+hr = HDLRegression(simulator="modelsim")
 # Add util, fw and VIP Scoreboard
 hr.add_files("../../../uvvm_util/src/*.vhd", "uvvm_util")
 hr.add_files("../../../uvvm_vvc_framework/src/*.vhd", "uvvm_vvc_framework")
@@ -41,7 +48,9 @@ hr.add_files("../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_v
 hr.add_files("../../tb/maintenance_tb/*.vhd", "bitvis_vip_i2c")
 # Add TB dependencies
 hr.add_files("../../tb/maintenance_tb/fpga_i2c_slave_github/*.vhd", "bitvis_vip_i2c")
-hr.add_files("../../tb/maintenance_tb/i2c_opencores/trunk/rtl/vhdl/*.vhd", "bitvis_vip_i2c")
+hr.add_files(
+    "../../tb/maintenance_tb/i2c_opencores/trunk/rtl/vhdl/*.vhd", "bitvis_vip_i2c"
+)
 
 # Add SBI VIP
 hr.add_files("../../../bitvis_vip_sbi/src/*.vhd", "bitvis_vip_sbi")
@@ -49,9 +58,20 @@ hr.add_files("../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_v
 
 # Add Wishbone VIP
 hr.add_files("../../../bitvis_vip_wishbone/src/*.vhd", "bitvis_vip_wishbone")
-hr.add_files("../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_vip_wishbone")
+hr.add_files(
+    "../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_vip_wishbone"
+)
 
-hr.start(regression_mode=True, gui_mode=False)
+sim_options = None
+default_options = []
+simulator_name = hr.settings.get_simulator_name()
+if simulator_name in ["MODELSIM", "RIVIERA"]:
+    sim_options = "-t ns"
+    # Set compile options
+    default_options = ["-suppress", "1346,1246,1236", "-2008"]
+    hr.set_simulator(simulator=simulator_name, com_options=default_options)
+
+hr.start(sim_options=sim_options)
 
 num_failing_tests = hr.get_num_fail_tests()
 num_passing_tests = hr.get_num_pass_tests()
@@ -61,6 +81,6 @@ if num_passing_tests == 0:
     sys.exit(1)
 # Remove output only if OK
 if hr.check_run_results(exp_fail=0) is True:
-    cleanup('Removing simulation output')
+    cleanup("Removing simulation output")
 # Return number of failing tests
 sys.exit(num_failing_tests)
